Cached UI Dispatcher in `ProgramRendererViewModel`
This resolves an issue where the GL thread is still running after the
application has been closed. When the application closes
`Current.Application` is null, so we can no longer access it to get the
Dispatcher. Caching the Dispatcher ensures that we have it available even
after the application closes.