Bug fixes to shader renaming
* Change renaming property on `ShaderListItemView` bind two ways
* Whenever the `ShaderListItemView` lost focus, it would confirm a new
name regardless. Since canceling also caused it to lose focus when the
name editing text box is hidden, this meant that canceling would also
confirm a new name, which is not the expected behavior. Added a check in
the lost focus variable to prevent this.
* Remove background color from `ShaderListItemView`
* Add check to rename shader command to prevent it from renaming the
shader if it is already being renamed

Pair programmed with @drbassett