Add active program
The active program will be the one that is rendered.

Program is set as active by selecting it in the program tree view, and
either double-clicking it, or selecting the 'Activate' option from the
context menu.

The currently active program is indicated with bold text in the program tree
view.