Change renaming system for programs
The old system placed a renaming field in all `ProgramViewModel`s, and
used it as the trigger for starting a renaming. Renaming was more a
property of the view than the view model, so it didn't really belong
there. Also, this system did not work at all if multiple views were bound
to the renaming property. All of them would activate at once, and
interfere with each other.

The new system internalizes renaming to the view, exposing a
command that can be bound. This system makes more sense intuitively, and
cooperates nicely with other views.