Update `ShaderCompiler` to handle multiple shaders
`ShaderCompiler` and a number of other classes from the proof-of-concept
were hard coded to handle only 1 program compesed of exactly 1 vertex
shader and 1 fragment shader. This commit backs out many of the features
from the proof-of-concept to remove this limitation and prepare Shader
Baker for future updates.

As mentioned in the summary, this commit also updates the original
`ShaderCompiler` class to validate any number of shaders. This commit also
modifies the GUI to display some of the information the shader compiler
produces, such as shader source code errors.