Add rudimentary serialization
All shaders and programs can be serialized into a file, and loaded at a
later time. This is primarily to save time spent testing the application.

The main GUI window accepts the shortcuts Ctrl+S to save a project, and
Ctrl+O to load it. The application always loads/saves to a file called
test.sbp in the working directory.

Also, the application now accepts a single command line argument to specify
the name of a project file to load at startup.